+++ CHANGELOG.md
@@ -11,7 +11,12 @@
   has been paused.
 - Fixes an issue where download errors would show a rev number instead of the Dropbox
   path.
+- Fixes a race condition when two processes try to start a sync daemon at the same time.
 - Fixes an issue in the macOS GUI where updating the displayed sync issues could fail.
+
+#### Removed:
+
+- Removed `lockfile` dependency.
 
 ## v1.0.2